split tests into correct files
